refactor: create RegisterForm for registration UI logic delete coments and unused imports
