+++ wealth-sync/src/components/Navbar.tsx
@@ -2,7 +2,6 @@ import LogoHomeButton from "./LogoHomeButton";
 import { ThemeModeToggle } from "./ThemeModeToggle";
 import NavbarLinks from "./NavbarLinks";
 import { NavbarAuth } from "./NavbarAuth";
-// import { NavbarAuth } from "./NavbarAuth";
 
 export function Navbar() {
   return (
